Add some casts in libveriuser to remove warnings.
The Cygwin compiler is a bit picky. This patch adds some casts
to remove compilation warnings. In the past I have had warnings
off because of problems with the STL, but for C directories
like this it makes sense to enable the warnings. It also does
not recognize that an assert(0) or assert(false) ends a routine
so it complains about no return at end of function or variables
not being defined.
(cherry picked from commit 6ae1f64c3cfa6eb75ad72c4674193eef7ceac37c)

+++ libveriuser/delay.c
@@ -43,7 +43,7 @@ int tf_isetdelay(PLI_INT32 delay, void*ss)
       if (pli_trace) {
 	    fprintf(pli_trace, "%s: tf_isetdelay(%d, ...)"
 		    " <unit=%d, prec=%d>;\n",
-		    vpi_get_str(vpiName, sys), delay, unit, prec);
+		    vpi_get_str(vpiName, sys), (int)delay, unit, prec);
       }
